feat:(Map): rename Map components to match LocationAutocomplete structure (#156)
BREAKING CHANGE: Map component is renamed to MapWithGoogleLoader and MapRenderer to Map to better reflect their function and to match the LocationAutocomplete API
