--- src/components/Map/MapRenderer/index.js
@@ -1,1 +1,0 @@
-export { default } from './MapRenderer';
