+++ src/components/Map/__tests__/MapWithGoogleLoader.spec.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import toJson from 'enzyme-to-json';
+import MapWithGoogleLoader from '../MapWithGoogleLoader';
+
+describe('<MapWithGoogleLoader/> that loads google api and renders a Map', () => {
+    it('should render with default props', () => {
+        const wrapper = mount(<MapWithGoogleLoader apiKey="someKey" />);
+
+        expect(toJson(wrapper)).toMatchSnapshot();
+    });
+});
