+++ src/index.js
@@ -10,7 +10,7 @@ export { default as Input } from './components/Input';
 export { default as Link } from './components/Link';
 export { List, ListItem, ListActions } from './components/List';
 export { default as Modal } from './components/Modal';
-export { Map, MapRenderer } from './components/Map';
+export { MapWithGoogleLoader, Map } from './components/Map';
 export { default as ProgressBar } from './components/ProgressBar';
 export { default as Slider } from './components/Slider';
 export { Tabs, Tab, TabContent, TabItem, TabMenu } from './components/Tabs';
